refactor: make CalculatorItemQueue<T> File
